# specs/001-boq-price-forecast/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: BoQ Price Forecast API
  version: 1.0.0
  description: |
    REST API for AI-powered BoQ price adjustment and forecasting.
    Supports Excel upload, weight attribution, price recalculation, forecasting, and defense document generation.
  contact:
    name: BoQ Price Forecast Team
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.boqprice.example.com
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
        params:
          type: object
      required: [detail]
...
    WeightAnalysisRequest:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        item_ids:
          type: array
          items:
            type: string
            format: uuid
        force_reanalyze:
          type: boolean
          default: false
      required: [job_id]
    WeightAnalysisResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
    WeightBreakdown:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
        code:
          type: string
        description_fa:
          type: string
        category:
          type: string
        llm_confidence:
          type: number
          format: double
        weights:
          type: array
          items:
            $ref: '#/components/schemas/ComponentWeight'
        fusion_confidence:
          type: number
          format: double
        updated_at:
          type: string
          format: date-time
    ComponentWeight:
      type: object
      properties:
        component_id:
          type: string
          format: uuid
        component_code:
          type: string
        component_name_fa:
          type: string
        llm_weight:
          type: number
          format: double
        ml_weight:
          type: number
          format: double
        final_weight:
          type: number
          format: double
        source:
          type: string
          enum: [llm, ml, fusion, expert]
        confidence:
          type: number
          format: double
        ml_r2:
          type: number
          format: double
        llm_certainty:
          type: number
          format: double
        overridden_by:
          type: string
          format: uuid
        overridden_at:
          type: string
          format: date-time
        override_reason:
          type: string
    WeightOverrideRequest:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
        component_id:
          type: string
          format: uuid
        new_weight:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
          minLength: 10
      required: [item_id, component_id, new_weight, reason]

    # ===== Price Calculation =====
...
paths:
...
  /weights/analyze:
    post:
      tags: [Weight Attribution]
      summary: Trigger AI weight attribution analysis
      operationId: analyzeWeights
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightAnalysisRequest'
      responses:
        '202':
          description: Analysis job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightAnalysisResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /weights/{job_id}:
    get:
      tags: [Weight Attribution]
      summary: Get weight breakdown for all items in job
      operationId: getWeightBreakdown
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Weight breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WeightBreakdown'
                  fusion_alpha:
                    type: number
                    format: double
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /weights/override:
    post:
      tags: [Weight Attribution]
      summary: Expert override of fused weight
      operationId: overrideWeight
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightOverrideRequest'
      responses:
        '200':
          description: Weight overridden successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentWeight'
        '400':
          description: Invalid weight or missing reason
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (requires reviewer/manager role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== Price Calculation =====
...